Backend_API/core_api/Serializers/direccion_serializer.py
from rest_framework import  serializers
from admin_tmaps.models import Direccion


class DireccionSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Direccion
        fields = ['empleado', 'latitud', 'longitud']